Move logic to retrieve system uuid to detect_utils
Being a generic helper function, get_uuid should be moved to
detect_utils.
This is also needed for the future refactor of detect_system.
