+++ hardware/detect_utils.py
@@ -16,6 +16,7 @@ import os
 import platform
 import subprocess
 import sys
+import uuid
 
 
 def cmd(cmdline):
@@ -231,3 +232,42 @@ def clean_str(val):
 def clean_tuples(lst):
     """Clean a list of tuples from bad UTF-8 strings."""
     return [tuple([clean_str(val) for val in elt]) for elt in lst]
+
+
+def _get_uuid_x86_64():
+    """Get uuid from dmidecode"""
+
+    uuid_cmd = subprocess.Popen(
+        "dmidecode -t 1 | grep UUID | awk '{print $2}'", shell=True,
+        stdout=subprocess.PIPE, universal_newlines=True)
+    stdout = uuid_cmd.communicate()[0]
+    return stdout.rstrip()
+
+
+def _get_uuid_ppc64le(hw_lst):
+    """Get uuid for ppc64 arch systems"""
+    vendor = None
+    serial = None
+    for (_, _, sys_subtype, value) in hw_lst:
+        if sys_subtype == 'vendor':
+            vendor = value
+        if sys_subtype == 'serial':
+            serial = value
+
+    system_uuid = None
+    system_uuid_fname = '/sys/firmware/devicetree/base/system-uuid'
+    if os.access(system_uuid_fname, os.R_OK):
+        with open(system_uuid_fname) as uuidfile:
+            system_uuid = uuidfile.read().rstrip(' \t\r\n\0')
+    elif vendor and serial:
+        root = uuid.UUID(bytes=b'\x00' * 16)
+        vendor_uuid = uuid.uuid5(root, vendor)
+        system_uuid = str(uuid.uuid5(vendor_uuid, serial))
+
+    return system_uuid
+
+
+def get_uuid(hw_lst):
+    if os.uname()[4] == 'ppc64le':
+        return _get_uuid_ppc64le(hw_lst)
+    return _get_uuid_x86_64()
